requirements.txt
pandas>=2.0.0
polars>=0.20.0
plotly>=5.18.0
streamlit>=1.30.0
requests>=2.31.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
Jinja2>=3.1.0
